Move DirReadJobQueue into DirTree.h
Despite the name, DirTree owns and controls the only instance of DirReadJobQueue.  Move the declaration into DirTree.h to reduce dependencies a little.
